refactor(services/pano-api): move proto out of service implementation
