+++ services/graphql/resolver/post_resolver.go
@@ -2,7 +2,7 @@ package resolver
 
 import (
 	"github.com/graph-gophers/graphql-go"
-	pano_api "go.kamp.us/services/pano-api/rpc/pano-api"
+	pano "go.kamp.us/protos/pano-api"
 )
 
 type postResolver struct {
