+++ services/graphql/resolver/resolver.go
@@ -1,6 +1,6 @@
 package resolver
 
-import pano_api "go.kamp.us/services/pano-api/rpc/pano-api"
+import pano "go.kamp.us/protos/pano-api"
 
 type Clients struct {
 	PanoAPI pano_api.PanoAPI
