+++ services/pano-api/server/get_comments.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/twitchtv/twirp"
-	api "go.kamp.us/services/pano-api/rpc/pano-api"
+	api "go.kamp.us/protos/pano-api"
 	"go.kamp.us/services/pano-api/server/helper"
 )
