+++ services/pano-api/server/is_upvoted.go
@@ -3,7 +3,7 @@ package server
 import (
 	"context"
 	"github.com/twitchtv/twirp"
-	api "go.kamp.us/services/pano-api/rpc/pano-api"
+	api "go.kamp.us/protos/pano-api"
 )
 
 func (s *PanoAPIServer) IsUpvoted(ctx context.Context, req *api.IsUpvotedRequest) (*api.IsUpvotedResponse, error) {
