+++ services/pano-api/server/update_comment.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/twitchtv/twirp"
-	api "go.kamp.us/services/pano-api/rpc/pano-api"
+	api "go.kamp.us/protos/pano-api"
 )
 
 func (s *PanoAPIServer) UpdateComment(ctx context.Context, req *api.UpdateCommentRequest) (*api.UpdateCommentResponse, error) {
